Added cookiestore to the supabase server call to provide auth context to the server,separated server and client components
